requirements/base.txt
# coding: utf-8

# -----------------------------------------------------------------------------

# Данный файл содержит общий для всех вариантов запуска список зависимостей

# проекта.

#

# Версии пакетов рекомендуется указывать не в виде конкретных номеров, а в виде

# диапазонов начиная с минимально необходимой для функционирования проекта

# версии заканчивая следующей мажорной версией, в которой могут появится

# обратно несовместимые изменения.

# 

# Это необходимо для того, чтобы избежать возможных конфликтов версий, когда в

# проекте указана одна версия, а в списке зависимостей какого-то другого пакета

# указана другая.

# -----------------------------------------------------------------------------


# Генерация документации.

Sphinx>=1.5.2

sphinx_rtd_theme>=0.1.9

sphinx-autobuild>=0.6.0



# Общесистемные низкоуровневые библиотеки.

pathlib2>=2.2.1,<3

psycopg2>=2.7.1,<3

# Django.

Django>=1.10.5,<1.11

django-mptt>=0.8.7,<1



# Пакеты платформы M3.

m3-core>=2.2.6,<2.3

m3-ui>=2.2.12,<2.3

m3-objectpack>=2.2.8,<2.3



# Вспомогательные библиотеки проектов.

educommon>=1.1.1,<2

wsfactory>=0.4.0,<0.5.0
